improve settings view

+++ NotoNote/ViewModels/SettingsViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using NotoNote.Models;
+using System.Collections.Specialized;
 
 namespace NotoNote.ViewModels;
 
@@ -17,15 +18,13 @@ public partial class SettingsViewModel : ObservableObject
         public bool Alt { get; set; }
     }
 
-    [ObservableProperty] private string _openAiApiKey;
+    [ObservableProperty] private string _openAiApiKey = string.Empty; // Initialize to avoid nullability issue
     [ObservableProperty] private Hotkey _hotkeyActivation;
     [ObservableProperty] private Hotkey _hotkeyToggleProfile;
-    private List<Profile> _profiles;
 
     public Keys[] AvailableKeys => Constants.AvailableKeys;
 
-    public List<Profile> Profiles => _profiles;
-
+    [ObservableProperty] List<Profile> _profiles = [];
     [ObservableProperty] Profile _selectedProfile;
 
     public SettingsViewModel(IProfileRepository profiles, IApiKeyRepository apiKey)
@@ -47,58 +46,75 @@ public partial class SettingsViewModel : ObservableObject
         var savedOpenAiApiKey = _apiKeyRepository.Get(ApiProvider.OpenAI);
         if (savedOpenAiApiKey != null) OpenAiApiKey = savedOpenAiApiKey.Value;
 
-        _profiles = _profilesRepository.GetAll().ToList();
+        _profiles = _profilesRepository.GetAll();
         _selectedProfile = _profiles[0];
     }
 
     partial void OnOpenAiApiKeyChanged(string value)
     {
-        if (string.IsNullOrEmpty(value))
-        {
-            _apiKeyRepository.Delete(ApiProvider.OpenAI);
-        }
-        else
-        {
-            var apiKey = new ApiKey(ApiProvider.OpenAI, value);
-            if (_apiKeyRepository.Get(ApiProvider.OpenAI) != null)
-            {
-                _apiKeyRepository.Update(apiKey);
-            }
-            else
-            {
-                _apiKeyRepository.Set(apiKey);
-            }
-        }
-    }
-
-    private void UpdateProfileRepository()
-    {
-        foreach (var profile in _profilesRepository.GetAll())
-        {
-
-        }
+        if (string.IsNullOrEmpty(value)) _apiKeyRepository.Delete(ApiProvider.OpenAI);
+        else _apiKeyRepository.AddOrUpdate(new ApiKey(ApiProvider.OpenAI, value));
     }
 
     [RelayCommand]
     private void AddProfile()
     {
-
+        var newProfile = new Profile(
+            new(Guid.NewGuid()),
+            new("NewProfile"),
+            new(""),
+            Constants.AvailableTranscriptionAiModels[0].Id,
+            Constants.AvailableChatAiModels[0].Id);
+        _profilesRepository.AddOrUpdate(newProfile);
+        Profiles = _profilesRepository.GetAll();
+        SelectedProfile = Profiles[^1];
     }
 
 
     [RelayCommand]
     private void DeleteProfile()
     {
+        if (SelectedProfile == null) return;
 
+        var allProfiles = _profilesRepository.GetAll();
+        if (allProfiles.Count == 1) return;
 
+        var index = allProfiles.IndexOf(SelectedProfile);
+
+        var isLast = index == allProfiles.Count - 1;
+        _profilesRepository.Delete(SelectedProfile.Id);
+
+        Profiles = _profilesRepository.GetAll();
+        SelectedProfile = Profiles[isLast ? index - 1 : index];
     }
 
 
     [RelayCommand]
-    private void MoveProfileToUp() { }
+    private void MoveProfileToUp()
+    {
+        if (SelectedProfile == null) return;
+        var allProfiles = _profilesRepository.GetAll();
+        if (allProfiles.Count == 1) return;
+        var index = allProfiles.IndexOf(SelectedProfile);
+        if (index == 0) return;
+        _profilesRepository.MoveIndex(SelectedProfile.Id, -1);
+
+        Profiles = _profilesRepository.GetAll();
+        SelectedProfile = Profiles[index - 1];
+    }
 
 
     [RelayCommand]
-    private void MoveProfileToDown() { }
+    private void MoveProfileToDown()
+    {
+        if (SelectedProfile == null) return;
+        var allProfiles = _profilesRepository.GetAll();
+        if (allProfiles.Count == 1) return;
+        var index = allProfiles.IndexOf(SelectedProfile);
+        if (index == allProfiles.Count - 1) return;
+        _profilesRepository.MoveIndex(SelectedProfile.Id, 1);
 
+        Profiles = _profilesRepository.GetAll();
+        SelectedProfile = Profiles[index + 1];
+    }
 }
